Moving all model traits to the concerns namespace
